+++ libcaf_core/test/flow/op/buffer.cpp
@@ -9,7 +9,7 @@
 #include "core-test.hpp"
 
 #include "caf/flow/coordinator.hpp"
-#include "caf/flow/item_publisher.hpp"
+#include "caf/flow/multicaster.hpp"
 #include "caf/flow/observable.hpp"
 #include "caf/flow/observable_builder.hpp"
 #include "caf/flow/observer.hpp"
@@ -118,7 +118,7 @@ SCENARIO("the buffer operator forces items at regular intervals") {
           cow_vector<int>{},        cow_vector<int>{64},
           cow_vector<int>{},        cow_vector<int>{128, 256, 512},
         };
-        auto pub = flow::item_publisher<int>{ctx.get()};
+        auto pub = flow::multicaster<int>{ctx.get()};
         sys.spawn([&pub, outputs](caf::event_based_actor* self) {
           pub.as_observable()
             .observe_on(self) //